# specs/001-company-role-user-setup/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GenzFeast — Company, Role, Category, Department & User Creation API
  version: "1.0.0"
  description: >
    Contract for feature 001-company-role-user-setup. All endpoints require
    Authorization: Bearer <JWT> (Architecture §4). company_id/role are always
    taken from the verified JWT claims, never from the request body or path.
    Every write additionally lands in audit_logs automatically via DB trigger
    (out of band for this contract).
servers:
  - url: https://api.genzfeast.example/v1
security:
  - bearerAuth: []
...
paths:
  /admin/companies:
    get:
      tags: [Companies]
      summary: List all companies (System Admin only)
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Company" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Companies]
      summary: Create a company (System Admin only). Roles are global and seeded once platform-wide, not per-company.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompanyCreateRequest" }
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/companies/{companyId}:
    patch:
      tags: [Companies]
      summary: Update company details or toggle open/closed status (System Admin only)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompanyUpdateRequest" }
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Companies]
      summary: >
        Soft-delete a company (System Admin only). Blocks every one of that
        company's users, in every role, from logging in ("blocked contact
        admin") and immediately revokes their active sessions — never a hard
        delete.
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "204": { description: Company soft-deleted; its users' sessions revoked }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/companies/{companyId}/admins:
    post:
      tags: [Users]
      summary: Create the first (or an additional) Company Admin for a company (System Admin only)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUserCreateRequest" }
      responses:
        "201":
          description: Company Admin created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
...
  /admin/categories/{categoryId}:
    patch:
      tags: [Categories]
      summary: Update a category belonging to any company (System Admin only)
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataUpdateRequest" }
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Categories]
      summary: Soft-delete a category belonging to any company (System Admin only)
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204": { description: Category soft-deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /tenant/categories:
    get:
      tags: [Categories]
      summary: >
        List categories for the caller's own company (Company Admin, Staff,
        Student). Server-side paginated (revised — this task's own explicit
        ask); `page`/`limit` both default when omitted, so an existing caller
        needing the full list (e.g. a dropdown) can still pass a high
        `limit` rather than paging through results.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: A page of the caller's company categories
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCategoriesResponse" }
    post:
      tags: [Categories]
      summary: Create a category scoped to the caller's own company (Company Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataCreateRequest" }
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/categories/{categoryId}:
    get:
      tags: [Categories]
      summary: >
        Not in this feature's original scope — added once `list` became
        paginated, so the edit screen can look up one category directly
        rather than fetch-the-full-list-and-find-by-id (Company Admin,
        Staff, Student).
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: The category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Categories]
      summary: Update a category scoped to the caller's own company (Company Admin only)
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterDataUpdateRequest" }
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Categories]
      summary: Soft-delete a category scoped to the caller's own company (Company Admin only)
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204": { description: Category soft-deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /auth/register/options:
    get:
      tags: [Users]
      summary: >
        Unauthenticated Category/Department lookup for the Registration
        screen's dropdowns (FR-012) — added alongside the mobile Login/
        Register wiring, since no other listing endpoint can be called
        before an account exists. Scoped by company_id the same way
        /auth/register itself is; id+name only, no other fields.
      parameters:
        - name: company_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The target company's active categories and departments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationOptions" }
  /auth/register:
    post:
      tags: [Users]
      summary: Student self-registration on a specific company's app
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentRegisterRequest" }
      responses:
        "201":
          description: Student account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409": { $ref: "#/components/responses/DuplicateUsername" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
    SearchParam:
      name: search
      in: query
      required: false
      description: Case-insensitive name substring filter
      schema: { type: string }
  responses:
    Forbidden:
      description: Caller's role/company scope does not permit this action
    NotFound:
      description: Resource does not exist or is not visible to the caller's scope
    ValidationError:
      description: Request failed field-level validation
    DuplicateUsername:
      description: Username already exists within the target company
  schemas:
    Company:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        contact_person: { type: string }
        mobile: { type: string }
        email: { type: string, format: email }
        address: { type: string }
        is_open: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CompanyCreateRequest:
      type: object
      required: [name, contact_person, mobile, email, address]
      properties:
        name: { type: string }
        contact_person: { type: string }
        mobile: { type: string }
        email: { type: string, format: email }
        address: { type: string }
    CompanyUpdateRequest:
      type: object
      properties:
        name: { type: string }
        contact_person: { type: string }
        mobile: { type: string }
        email: { type: string, format: email }
        address: { type: string }
        is_open: { type: boolean }

    # Global, platform-wide catalog — not per-Company (revised; see
    # migration 20260907170000_global_roles and 20260907180000).
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        company_id: { type: string, format: uuid }
        name: { type: string, example: "Student" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaginatedCategoriesResponse:
      type: object
      description: Server-side pagination (10 per page) — the mobile list screen fetches the next page on scroll.
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    MasterDataCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    MasterDataUpdateRequest:
      type: object
      properties:
        name: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        company_id: { type: string, format: uuid, nullable: true }
        role: { type: string, enum: [system_admin, company_admin, company_staff, student, teaching, non_teaching] }
        name: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender], nullable: true }
        category_id: { type: string, format: uuid, nullable: true }
        department_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [active, inactive, locked] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AdminUserCreateRequest:
      type: object
      required: [name, username, password, email]
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password, description: "Temporary password" }
        email: { type: string, format: email }
...
    StudentRegisterRequest:
      type: object
      required: [name, username, password, email, gender, category_id, company_id]
      properties:
        name: { type: string }
        username: { type: string, description: "Mobile number" }
        password: { type: string, format: password }
        email: { type: string, format: email }
        gender: { type: string, enum: [male, female, transgender] }
        category_id: { type: string, format: uuid }
        department_id: { type: string, format: uuid, nullable: true }
        company_id:
          type: string
          format: uuid
          description: >
            Not in this endpoint's original schema — added for consistency
            with the identical company_id-per-branded-build convention
            002/003 established for every other unauthenticated /auth/*
            endpoint.
    RegistrationOption:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    RegistrationOptions:
      type: object
      properties:
        categories:
          type: array
          items: { $ref: "#/components/schemas/RegistrationOption" }
        departments:
          type: array
          items: { $ref: "#/components/schemas/RegistrationOption" }